use sodiumoxide over libsodium_ffi, rust 2018
fixes #3

this commit updates us to rust 2018 epoch! no more extern crates \o/
this commit also removes our custom libsodium wrapper (which was
causing errors like #3), and uses a more polished solution.
sodiumoxide. as a result of this we now have 0 unsafe code.

I'm sure we can clean up the interface now that we're using sodium
oxide, but I'd prefer to wait, as to not make a breaking change
to the interface.