Remove mut keyword from key parameters in v2/local.rs
It's unclear why this was mut, so I may be overlooking something.  That
being said, the use of a mutable slice for what seemed a non-returning
constant parameter appeared a code smell.  By avoiding unnecessary
mutability, we may sidestep a class of accidental errors and hopefully
have more robust (and correct) software.